@@ -1,6 +1,6 @@
 module github.com/rancher-sandbox/rancher-desktop/src/go/networking
 
-go 1.23.0
+go 1.24.0
 
 toolchain go1.24.1
 
